feature: add a trash icon to owner's cards only

+++ src/components/Card.js
@@ -1,15 +1,17 @@
 
 export default class Card{
-    constructor(data, templateSelector, {handleCardClick, handleDeleteCard, handleLikeClick}){
+    constructor(data, idOwner, templateSelector, {handleCardClick, handleDeleteCard, handleLikeClick}){
         this._name = data.name;
         this._link = data.link;
         this._likes = data.likes;
         this._id = data._id;
         this._owner = data.owner;
+        this._idOwner = idOwner;
         this._templateSelector = templateSelector; 
         //функция открытия попапа с картинкой по клику на карточку
         this._handleCardClick = handleCardClick; 
         this._handleLikeClick = handleLikeClick;
+        this._handleDeleteCard = handleDeleteCard;
     }
 
     _getTemplate() {
@@ -46,29 +48,35 @@ export default class Card{
         this._cardImage.addEventListener('click', () => { 
             this._handleCardClick(this._name, this._link); 
         });
-        this._deleteButton.addEventListener('click', () => {
-             this._deleteElement(); 
-        });
     }
 
     getCard(){
         return this;
     }
 
+    //отрисовать иконку "удалить" только у карточек владельца
+    _renderTrashIcon(){
+        if (this._idOwner === this._owner._id) {
+            this._deleteButton = this._element.querySelector('.card__delete');
+            this._deleteButton.classList.add('card__delete_active');
+            this._deleteButton.addEventListener('click', () => {
+                this._deleteElement();
+            });
+        }
+    }
+
     generateCard(){
         this._element = this._getTemplate();
         this._likeButton = this._element.querySelector('.card__like');
         this._cardImage = this._element.querySelector('.card__image');
-        this._deleteButton = this._element.querySelector('.card__delete');
-
-        this._setEventListener();
 
         this._element.querySelector('.card__title').textContent = this._name;
         this._cardImage.src = this._link;
         this._cardImage.alt = "Фото пользователя " + this._name;
         
         this.updateCounterLikes(this._likes.length);
-
+        this._renderTrashIcon();
+        this._setEventListener();
         return this._element;
     }
 }
